assets of first page and icons added, a doc called first page added,
The HTML Doc called First page needs work, css was kept internal so fixing it with AI would be easier. still needs work and more color, the mascot and the comment at the end of body will be added, button layout will be removed. (changing the buttons with on-click images can help)